# Move bytes.buffer to []byte for set and message