--- a/pkg/entities/message.go
+++ b/pkg/entities/message.go
@@ -15,7 +15,6 @@
 package entities
 
 import (
-	"bytes"
 	"encoding/binary"
 )
 
@@ -23,13 +22,14 @@ const (
 	MaxTcpSocketMsgSize int = 65535
 	DefaultUDPMsgSize   int = 512
 	MaxUDPMsgSize       int = 1500
+	MsgHeaderLength     int = 16
 )
 
 // Message represents IPFIX message.
 // TODO: Currently, it supports only one set. This will be extended to support multiple
 // sets.
 type Message struct {
-	buffer        *bytes.Buffer
+	msgHeader     []byte
 	version       uint16
 	length        uint16
 	seqNumber     uint32
@@ -42,7 +42,7 @@ type Message struct {
 
 func NewMessage(isDecoding bool) *Message {
 	return &Message{
-		buffer:     &bytes.Buffer{},
+		msgHeader:  make([]byte, MsgHeaderLength),
 		isDecoding: isDecoding,
 	}
 }
@@ -54,7 +54,7 @@ func (m *Message) GetVersion() uint16 {
 func (m *Message) SetVersion(version uint16) {
 	m.version = version
 	if !m.isDecoding {
-		binary.BigEndian.PutUint16(m.buffer.Bytes()[0:2], version)
+		binary.BigEndian.PutUint16(m.msgHeader[0:2], version)
 	}
 }
 
@@ -65,7 +65,7 @@ func (m *Message) GetMessageLen() uint16 {
 func (m *Message) SetMessageLen(len uint16) {
 	m.length = len
 	if !m.isDecoding {
-		binary.BigEndian.PutUint16(m.buffer.Bytes()[2:4], len)
+		binary.BigEndian.PutUint16(m.msgHeader[2:4], len)
 	}
 }
 
@@ -76,7 +76,7 @@ func (m *Message) GetSequenceNum() uint32 {
 func (m *Message) SetSequenceNum(seqNum uint32) {
 	m.seqNumber = seqNum
 	if !m.isDecoding {
-		binary.BigEndian.PutUint32(m.buffer.Bytes()[8:12], seqNum)
+		binary.BigEndian.PutUint32(m.msgHeader[8:12], seqNum)
 	}
 }
 
@@ -87,7 +87,7 @@ func (m *Message) GetObsDomainID() uint32 {
 func (m *Message) SetObsDomainID(obsDomainID uint32) {
 	m.obsDomainID = obsDomainID
 	if !m.isDecoding {
-		binary.BigEndian.PutUint32(m.buffer.Bytes()[12:], obsDomainID)
+		binary.BigEndian.PutUint32(m.msgHeader[12:], obsDomainID)
 	}
 }
 
@@ -98,7 +98,7 @@ func (m *Message) GetExportTime() uint32 {
 func (m *Message) SetExportTime(exportTime uint32) {
 	m.exportTime = exportTime
 	if !m.isDecoding {
-		binary.BigEndian.PutUint32(m.buffer.Bytes()[4:8], exportTime)
+		binary.BigEndian.PutUint32(m.msgHeader[4:8], exportTime)
 	}
 }
 
@@ -118,23 +118,11 @@ func (m *Message) AddSet(set Set) {
 	m.set = set
 }
 
-func (m *Message) GetMsgBuffer() *bytes.Buffer {
-	return m.buffer
+func (m *Message) GetMsgHeader() []byte {
+	return m.msgHeader
 }
 
-func (m *Message) GetMsgBufferLen() int {
-	return m.buffer.Len()
-}
-
-func (m *Message) WriteToMsgBuffer(bytesToWrite []byte) (int, error) {
-	return m.buffer.Write(bytesToWrite)
-}
-
-func (m *Message) CreateHeader() (int, error) {
-	header := make([]byte, 16)
-	return m.WriteToMsgBuffer(header)
-}
-
-func (m *Message) ResetMsgBuffer() {
-	m.buffer.Reset()
+func (m *Message) ResetMsgHeader() {
+	m.msgHeader = nil
+	m.msgHeader = make([]byte, MsgHeaderLength)
 }