This PR is about modifying `bytes.Buffer` struct to `[]byte` for set and message. It also optimizes the way of encoding record, set and message to reduce memory allocs.
Before
```
BenchmarkExportConntrackConns-2   	     154	   9073336 ns/op	  378045 B/op	   15034 allocs/op
BenchmarkExportDenyConns-2   	     252	   5222296 ns/op	  231724 B/op	    9411 allocs/op
```
After
```
BenchmarkExportConntrackConns-2   	     168	   7869668 ns/op	  297198 B/op	    5651 allocs/op
BenchmarkExportDenyConns-2   	     259	   5044816 ns/op	  183199 B/op	    3385 allocs/op
```
For conntrack connections exporting, CPU time gets reduced by 13.27%, memory usage reduced by 21.39%, allocs reduced by 62.41%. 
For deny connections exporting, CPU time gets reduced by 3.40%, memory usage reduced by 20.94%, allocs reduced by 64.03%