@zyiou 2021-06-16
I tried this change with Flow Exporter performance benchmarking but unfortunately, didn't see any improvement. Wondering whether we need to keep this PR or close it. Pros I can think of keeping is that it makes encoding of record, set, message consistent with []byte. @srikartati Please also have a try on your side for us to decide. Thanks!