@srikartati 2021-06-17
Here is the comparison in my runs:
Before fix:
```
vagrant@k8s-node-worker-1:~/antrea/pkg/flowaggregator$ go test -test.v -run=none -test.benchmem  -bench=. -count=2 -memprofile memprofile.out -cpuprofile profile.out 
goos: linux
goarch: amd64
pkg: antrea.io/antrea/pkg/flowaggregator
cpu: Intel(R) Core(TM) i9-9980HK CPU @ 2.40GHz
BenchmarkIntraNodeFlowRecords
    flowaggregator_perf_test.go:458: Num messages received: 1550264
BenchmarkIntraNodeFlowRecords-2   	       1	120004095225 ns/op	8876354472 B/op	332625061 allocs/op
PASS
```

After fix:
```
vagrant@k8s-node-worker-1:~/antrea/pkg/flowaggregator$ go test -test.v -run=none -test.benchmem  -bench=. -memprofile memprofile.out -cpuprofile profile.out 
goos: linux
goarch: amd64
pkg: antrea.io/antrea/pkg/flowaggregator
cpu: Intel(R) Core(TM) i9-9980HK CPU @ 2.40GHz
BenchmarkIntraNodeFlowRecords
    flowaggregator_perf_test.go:479: Num messages received: 1536208
BenchmarkIntraNodeFlowRecords-2   	       1	120002065282 ns/op	9408062880 B/op	340186942 allocs/op
PASS
ok  	antrea.io/antrea/pkg/flowaggregator	120.304s
```

There is no real improvement. 
As mentioned offline, there are different entities (record, set, message) and we are allocating memory for bytes at each entity. We should think of a way to read the memory at each entity rather than writing them again and again.
We should look at chaining readers like [here](https://medium.com/learning-the-go-programming-language/streaming-io-in-go-d93507931185)?