@zyiou 2021-06-24
> I like the approach of maintaining all element lengths and then encoding. The encoding function can be futrther enhanced by taking the allocated bytes and filling them. We can reduce the allocations in the encoding function. What do you think?

Yes, but we will need to keep element list instead which is not necessary for encoding. Tested with latest change with pre-allocation for set and record. But the result seems not improving:
Before:
```
BenchmarkPoll-2  	  112	  9965818 ns/op	 902472 B/op	  54534 allocs/op
BenchmarkExportConntrackConns-2  106	  9684719 ns/op	 685126 B/op	  15782 allocs/op
BenchmarkExportDenyConns-2  	  247	  5021390 ns/op	 215250 B/op	  6539 allocs/op

```
After
```
BenchmarkPoll-2  	  127	  8602612 ns/op	 887633 B/op	  54184 allocs/op
BenchmarkExportConntrackConns-2  123	  8935159 ns/op	 679445 B/op	  18822 allocs/op
BenchmarkExportDenyConns-2  	  204	  4964911 ns/op	 304444 B/op	  10967 allocs/op
```