@srikartati 2021-06-29
> Yes, but we will need to keep element list instead which is not necessary for encoding. Tested with latest change with pre-allocation for set and record. But the result seems not improving:
> Before:
> 
> ```
> BenchmarkPoll-2  	  112	  9965818 ns/op	 902472 B/op	  54534 allocs/op
> BenchmarkExportConntrackConns-2  106	  9684719 ns/op	 685126 B/op	  15782 allocs/op
> BenchmarkExportDenyConns-2  	  247	  5021390 ns/op	 215250 B/op	  6539 allocs/op
> ```
> 
> After
> 
> ```
> BenchmarkPoll-2  	  127	  8602612 ns/op	 887633 B/op	  54184 allocs/op
> BenchmarkExportConntrackConns-2  123	  8935159 ns/op	 679445 B/op	  18822 allocs/op
> BenchmarkExportDenyConns-2  	  204	  4964911 ns/op	 304444 B/op	  10967 allocs/op
> ```

CPU time seems to be better even though allocations have gone up.
I looked at the change. We are still writing to a slice at all entities: record, set and message. Once we write to bytes as part of encoding.. could we just read them in other entities. Can this alternative work?