@zyiou 2021-07-01
> CPU time seems to be better even though allocations have gone up.
> I looked at the change. We are still writing to a slice at all entities: record, set and message. Once we write to bytes as part of encoding.. could we just read them in other entities. Can this alternative work?

Yes, thanks for the suggestion. I removed the write to set (record and message will need the write for storing record and send message respectively) and saw some improvement in CPU and memory:
Before
```
BenchmarkExportConntrackConns-2   	     154	   9073336 ns/op	  378045 B/op	   15034 allocs/op
BenchmarkExportDenyConns-2   	     252	   5222296 ns/op	  231724 B/op	    9411 allocs/op
```
After
```
BenchmarkExportConntrackConns-2   	     187	   8077640 ns/op	  317389 B/op	   12054 allocs/op
BenchmarkExportDenyConns-2   	     301	   4548239 ns/op	  188549 B/op	    7381 allocs/op
```