@zyiou 2021-07-06
> One more thing with `EncodeToIEDataType`. We are allocating bytes two times. Once for the record and doing it again inside `EncodeToIEDataType`. Could we send the pointer of allocated memory and fill it inside the Encode function?

Sure. After sending allocated memory to encoding function, alloc gets great improvement :)
Before
```
BenchmarkExportConntrackConns-2   	     187	   8077640 ns/op	  317389 B/op	   12054 allocs/op
BenchmarkExportDenyConns-2   	     301	   4548239 ns/op	  188549 B/op	    7381 allocs/op
```
After
```
BenchmarkExportConntrackConns-2   	     168	   7869668 ns/op	  297198 B/op	    5651 allocs/op
BenchmarkExportDenyConns-2   	     259	   5044816 ns/op	  183199 B/op	    3385 allocs/op
```